fix in user

+++ src/main/webapp/WEB-INF/view/user.jsp
@@ -3,6 +3,7 @@
 <head>
 <script>
 	var tbUser;
+	var modeSubmit = 'insert';
 	
 	function getFormData($form) {
 		var unindexed_array = $form.serializeArray();
@@ -19,7 +20,7 @@
 			type : 'get',
 			url : 'role/',
 			success : function(d) {
-				var s = '<option">Choose Role</option>';
+				var s = '<option>Choose Role</option>';
 
 				$(d).each(function(index, element) {
 					s += '<option value="' + element.id + '" data-nama="' + element.name + '">'
@@ -46,24 +47,26 @@
 			success : function(d) {
 				tbUser.clear().draw();
 				$(d).each(function(index, element) {
-					tbUser.row.add([
-						element.username,
-						element.role.name,
-						element.email,
-						'<div class="dropdown"><button class="btn btn-default dropdown-toggle" type="button" data-toggle="dropdown">'
-						+ '<i class="fa fa-bars"></i></button><ul class="dropdown-menu">'
-						+ '<li><a href="javascript:void(0)" onclick="loadData('
-						+ element.id
-						+ ')">Edit</a></li>'
-						+ '<li><a href="javascript:void(0)" onclick="loadReset('
-						+ element.id
-						+ ')">Reset Password</a></li>'
-						+ '<li><a href="javascript:void(0)" onclick="deleteDisabled('
-						+ element.id
-						+ ')">Delete</a></li>'
-						+ '</ul></div>',
-						element.id ])
-					.draw();
+					if (element.deleted == false) {
+						tbUser.row.add([
+							element.username,
+							element.role.name,
+							element.email,
+							'<div class="dropdown"><button class="btn btn-default dropdown-toggle" type="button" data-toggle="dropdown">'
+							+ '<i class="fa fa-bars"></i></button><ul class="dropdown-menu">'
+							+ '<li><a href="javascript:void(0)" data-toggle="modal" data-target="#modal-edit" onclick="loadData('
+							+ element.id
+							+ ')">Edit</a></li>'
+							+ '<li><a href="javascript:void(0)" data-toggle="modal" data-target="#modal-reset" onclick="loadData('
+							+ element.id
+							+ ')">Reset Password</a></li>'
+							+ '<li><a href="javascript:void(0)" onclick="deleteDisabled('
+							+ element.id
+							+ ')">Delete</a></li>'
+							+ '</ul></div>',
+							element.id ])
+						.draw();
+					}
 				});
 			},
 		});
@@ -71,56 +74,43 @@
 
 	// function untuk load edit data user
 	function loadData(id) {
+		reset();
 		$.ajax({
-			type : 'get',
-			url : 'user/' + id,
-			success : function(d) {
-				$('#form-edit input[name=idEdit]').val(d.id);
-				$('select[name=roleId2]').val(d.roleId);
-				$('input[name=email]').val(d.email);
-				$('input[name=username]').val(d.username);
-				$('input[name=mobileFlag]').val(d.mobileFlag);
-				$('input[name=mobileToken]').val(d.mobileToken);
-				$('#form-edit').trigger('reset');
-			},
-			error : function(d) {
-				console.log('Error!');
-			}
-		});
-		$('#modal-edit').modal('show');
-	}
-
-	// function untuk load reset data user
-	function loadReset(id) {
-		$.ajax({
-			type : 'get',
-			url : 'user/' + id,
-			success : function(d) {
-				$('#form-reset input[name=idEdit]').val(d.id);
-				$('input[name=password]').val(d.password);
-				$('#form-reset').trigger('reset');
-			},
-			error : function(d) {
-				console.log('Error!');
-			}
-		});
-		$('#modal-reset').modal('show');
+				type : 'get',
+				url : 'user/' + id,
+				success : function(d) {
+					$('input[name=id]').val(d.id);
+					$('select[name=roleId]').val(d.roleId);
+					$('input[name=roleId]').val(d.roleId);
+					$('input[name=email]').val(d.email);
+					$('input[name=username]').val(d.username);
+					$('#form-edit input[name=password]').val(d.password);
+					$('input[name=mobileFlag]').val([d.mobileFlag]);
+					$('input[name=mobileToken]').val(d.mobileToken);
+					modeSubmit ='update';
+				},
+				error : function(d) {
+					console.log('Error!');
+				}
+			});
 	}
 	
 	// function untuk save data user
 	function saveData(tipe) {
 		var method;
 		if (tipe == 'add') {
-			method = 'POST';
 			var data = getFormData($('#form-add'));
+			$('#modal-add').modal('hide');
+			method = 'POST';
 		} else if (tipe == 'edit') {
-			method = 'PUT';
 			var data = getFormData($('#form-edit'));
-		} else if (tipe == 'reset') {
+			$('#modal-edit').modal('hide');
 			method = 'PUT';
+		} else if (tipe == 'reset') {
 			var data = getFormData($('#form-reset'));
+			$('#modal-reset').modal('hide');
+			method = 'PUT';
 		}
-
 		$.ajax({
 			type : method,
 			url : 'user/',
@@ -128,12 +118,12 @@
 			contentType : 'application/json',
 			success : function(d) {
 				showData();
-				$('#modal-add').modal('hide');
-				$('#modal-edit').modal('hide');
-				$('#modal-reset').modal('hide');
-				$('#form-add').modal('reset');
-				$('#form-edit').modal('reset');
-				$('#form-reset').modal('reset');
+				if (d.retypePass == d.password) {
+					$('#form-reset input[name=password]').val(d.password);
+				} else {
+					alert("Password tidak sama.\nKetik ulang password Anda!");
+				}
+				modeSubmit = 'insert';
 			},
 			error : function(d) {
 				console.log('Error');
@@ -155,6 +145,16 @@
 				}
 			});
 		}
+	}
+
+	// function untuk reset
+	function reset() {
+		$('#form-add').trigger('reset');
+		$('#form-edit').trigger('reset');
+		$('#form-reset').trigger('reset');
+		$('#form-add input[type=hidden]').val('');
+		$('#form-edit input[type=hidden]').val('');
+		$('#form-reset input[type=hidden]').val('');
 	}
 
 	$(document).ready(function() {
@@ -186,7 +186,7 @@
 			</div>
 			<div class="col-xs-5">
 				<button type="button" class="btn btn-default pull-right hidden-xs"
-					data-toggle="modal" data-target="#modal-add"
+					data-toggle="modal" data-target="#modal-add" onclick="reset()"
 					style="width: 30px; height: 30px; padding: 0px 0px">
 					<i class="fa fa-plus" style="font-size: 20px"></i>
 				</button>
@@ -309,11 +309,15 @@
 												class="form-control" name="username" id="username">
 										</div>
 										<div class="form-group">
+											<input type="hidden"
+												class="form-control" name="password" id="password">
+										</div>
+										<div class="form-group">
 											<label>Mobile Flag</label>
 											<div class="input-group">
 												<input type="radio" name="mobileFlag" id="mobileFlag"
-													value=true> True &nbsp; <input type="radio"
-													name="mobileFlag" id="mobileFlag" value=false>
+													value="true"> True &nbsp; <input type="radio"
+													name="mobileFlag" id="mobileFlag" value="false">
 												False
 											</div>
 										</div>
@@ -356,15 +360,37 @@
 								<div class="row">
 									<div class="col-xs-12">
 										<div class="form-group">
-											<input type="hidden" class="form-control" name="id" id="idEdit">
+											<input type="text" class="form-control" name="id" id="idEdit">
 										</div>
 										<div class="form-group">
-											<label>Password</label> <input type="text"
+											<input type="text" class="form-control"
+												id="roleId2" name="roleId">
+										</div>
+										<div class="form-group">
+											<input type="hidden" class="form-control"
+												name="email" id="email">
+										</div>
+										<div class="form-group">
+											<input type="hidden"
+												class="form-control" name="username" id="username">
+										</div>
+										<div class="form-group">
+											<label>Password</label><input type="text"
 												class="form-control" name="password" id="password">
 										</div>
 										<div class="form-group">
 											<label>Re-type Password</label> <input type="text"
 												class="form-control" name="retypePass" id="retypePass">
+										</div>
+										<div class="form-group">
+											<div class="input-group">
+												<input type="hidden" class="form-control" name="mobileFlag"
+													id="mobileFlag">
+											</div>
+										</div>
+										<div class="form-group">
+											<input type="hidden"
+												class="form-control" name="mobileToken" id="mobileToken">
 										</div>
 									</div>
 								</div>
